requirements.txt
Flask
transformers
torch
readability-lxml
gunicorn
